Update Version (#5)
Update Typer Dependencies along with removing the access time check for
file expiry.

If we are checking for file expiry every week, then we are accessing it
each week. Therefore, it would never actually correctly check.

Update version (#6)

Update version and add check into file expiry files for the last date
for which a file is considered 'expired'.

Machung/removeprint (#7)